pyproject.toml
[tool.poetry]
name = "adventofcode"
version = "0.1.0"
description = "To solve puzzles"
authors = ["Viktor"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "aoclib"},
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
numpy = "^1.23.5"
matplotlib = "^3.6.2"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
